requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.2
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-filter==23.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-core==4.4.1
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.2
urllib3==1.26.15
